integrations/m365-copilot/requirements.txt
azure-identity==1.15.0
azure-core==1.29.5
requests==2.31.0
